We're linting, linting in the name of the lord

+++ internal/pkg/consensus/validation_test.go
@@ -153,7 +153,7 @@ func TestIngestionValidator(t *testing.T) {
 		require.NoError(t, err)
 		assert.NoError(t, validator.Validate(ctx, msg))
 
-		highNonce := uint64(act.CallSeqNum + mpoolCfg.MaxNonceGap + 10)
+		highNonce := act.CallSeqNum + mpoolCfg.MaxNonceGap + 10
 		msg, err = types.NewSignedMessage(*newMessage(t, alice, bob, highNonce, 5, 1, 0), signer)
 		require.NoError(t, err)
 		err = validator.Validate(ctx, msg)
